vim: vimrc.d: add FontSize command
GVim is completely ridiculous when it comes to changing the font size quickly
on demand.  This is particularly awkward and embarrassing when I am teaching,
live projecting, or otherwise demoing something in GVim on a projector with
a live audience.

This adds the FontSize(size) function and the associated FontSize command
(to save a little typing).  Used like this:

:FontSize 10
:FontSize 20

Makes an insane amount of sense, doesn't it...