vim: after: fix spell checking in latex files
This one was a real pain in the ass.  There seems to be an issue with
/usr/share/vim/vim74/syntax/tex.vim

Analyzing :scriptnames shows that if
syntax spell toplevel

is set before /usr/share/vim/vim74/syntax/tex.vim loads, the command
has no effect.  Putting it in after/plugin didn't work.  Even
putting it in after/syntax/plaintex.vim didn't work.

Seems that setting this in after/syntax/tex.vim is the /only/
solution.  What the hell.