vim: YouCompleteMe: add --clangd-completer
Apparently I had not updated YouCompleteMe in /forever/.  Perhaps it was
the update, perhaps it was the move from install.sh to install.py (which
I assume it was), but now I have to explicitly enable support for C?!

Fine.

The new clangd driven semantic completion is a lot louder and more
annoying than the old libclang implementation. ::sigh::